4_Implementacion/backend/requirements.txt
flask
flask_pymongo
python-dotenv
pymongo
flask-restful
flask-jwt-extended
certifi